src/langchain_agent/requirements.txt
# LangChain Agent 依赖

# LangChain 1.0 核心
langchain>=1.0.0
langchain-core>=1.0.0

# LLM提供商（使用langchain-openai连接智谱AI）
langchain-openai>=1.0.0

# MCP集成
mcp>=1.0.0

# 向量存储
langchain-chroma>=1.0.0
chromadb>=0.5.0

# WebUI
streamlit>=1.30.0

# 配置管理
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# 日志
loguru>=0.7.0

# HTTP请求
httpx>=0.25.0
aiohttp>=3.9.0

# 类型提示
typing-extensions>=4.8.0

